chore: relocate navbar to components
