+++ website-frontend/src/lib/components/nav/NavBar.svelte
